rename diploid.hpp
