+++ fwdpy11/headers/fwdpy11/types.hpp
@@ -28,7 +28,7 @@
 #include <pybind11/pybind11.h>
 //#include <fwdpy11/opaque/opaque_types.hpp>
 #include "types/typedefs.hpp"
-#include "types/diploid.hpp"
+#include "types/Diploid.hpp"
 #include <fwdpy11/rng.hpp>
 #include <fwdpp/sugar/popgenmut.hpp>
 //#include <fwdpp/sugar/generalmut.hpp>
